requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
pandas
